Refactored connecting, and reconnecting
The user either logs in with an account, in which case its name, and the
password is kept for reconnects. Or the user can login with a jid (and
server) which does not currently exist as a local account.  On success
an account is created, with the jid as the name.