Adding plugin interface for C files.
Implementation of hooks using C interface basedon dlfcn.  Added test
plugin and makefile to build it.  In order to test it add the followin
into your profrc:
    [plugins]
    load=test-c-plugin.so
and execute profanity piping stderr to some file.  The file should
contain all entries whenever the plugin function is triggered.  It seem
to be workin but some parts are missing.