autoconf adjustments for OpenBSD
In OpenBSD we have two readline libs. On the one hand we have an old
version in the base OS and on the other hand we have an new one in the
ports tree. profanity needs the new one and should pick up the right
one. readline from ports tree is prefixed as ereadline.

Note:
Original patch by Rafael Sadowski.
Cleaned up by jubalh with heavy help of pasis.