Remove double declaration and reduce scope of pcontact
In command/commands.c you declare pcontact at the top of the function.
Later in one case you declare it again and use it in a small scope. And
in the same function you use it again this time from the first
declaration. I think you intended only declaring it one time. At first I
thought about top of the function because its easier to find. Then I saw
that you declare "Jid jid" not at the top too, and I think for the sake
of reducing scope it is better. So I went for this approach.