Remove not needed initialization in chatwin_new()
00dbc1ba7c66c3bb6898aed26922bf5295768bb1 attempted to fix
https://github.com/profanity-im/profanity/issues/1449.

Here we revert the part about initializing is_omemo in chatwin_new()
since this is done in win_create_chat() already.

The actual mistake was the wrong cast of the muc window which is fixed
by the same commit. The mistake was introduced in in 3370418d71de255c832da97113543e554ec0e86b.

While being at it we also remove the is_ox initialization since this is
handled in win_create_chat()/win_create_muc() too.