Add Jingle: XEP-0166, XEP-0234, XEP-0261, XEP-0047
Jingle is a session negotiation protocol,
it can not transfer data by itself,
but it enables us to transfer data
over negotiated channels.

To transfer files over Jingle, the IBB is a required channel,
it's very stable (over XMPP), but it's slow.

It's planned to add Socks5 and over protocols in the future
updates,
but for now this is a minimal setup to use the Jingle for file transfer.

Special thanks to the @jubalh and @jaeckel for their support and
contributions.